+++ src/main/java/blusunrize/immersiveengineering/client/render/tile/BucketWheelRenderer.java
@@ -10,41 +10,47 @@ package blusunrize.immersiveengineering.client.render.tile;
 
 import blusunrize.immersiveengineering.api.IEProperties.IEObjState;
 import blusunrize.immersiveengineering.api.IEProperties.VisibilityList;
-import blusunrize.immersiveengineering.client.ClientUtils;
-import blusunrize.immersiveengineering.client.models.IOBJModelCallback;
+import blusunrize.immersiveengineering.api.client.IVertexBufferHolder;
 import blusunrize.immersiveengineering.client.models.obj.IESmartObjModel;
-import blusunrize.immersiveengineering.client.utils.SinglePropertyModelData;
 import blusunrize.immersiveengineering.common.blocks.IEBlocks.Multiblocks;
 import blusunrize.immersiveengineering.common.blocks.metal.BucketWheelTileEntity;
 import blusunrize.immersiveengineering.common.config.IEServerConfig;
 import blusunrize.immersiveengineering.common.util.Utils;
+import com.google.common.cache.Cache;
+import com.google.common.cache.CacheBuilder;
 import com.google.common.collect.Lists;
 import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.vertex.IVertexBuilder;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.BlockRendererDispatcher;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
 import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.model.BakedQuad;
 import net.minecraft.client.renderer.model.IBakedModel;
 import net.minecraft.client.renderer.tileentity.TileEntityRenderer;
 import net.minecraft.client.renderer.tileentity.TileEntityRendererDispatcher;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.Direction;
 import net.minecraft.util.math.vector.Quaternion;
 import net.minecraft.util.math.vector.Vector3f;
-import net.minecraftforge.client.model.data.IModelData;
+import net.minecraftforge.client.model.data.EmptyModelData;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
 
 @SuppressWarnings("deprecation")
 public class BucketWheelRenderer extends TileEntityRenderer<BucketWheelTileEntity>
 {
 	public static DynamicModel<Void> WHEEL;
+	private static final Cache<List<String>, IVertexBufferHolder> CACHED_BUFFERS = CacheBuilder.newBuilder()
+			.maximumSize(100)
+			.expireAfterAccess(1, TimeUnit.MINUTES)
+			.<Object, IVertexBufferHolder>removalListener(rem -> rem.getValue().reset())
+			.build();
 
 	public BucketWheelRenderer(TileEntityRendererDispatcher rendererDispatcherIn)
 	{
@@ -56,47 +62,66 @@ public class BucketWheelRenderer extends TileEntityRenderer<BucketWheelTileEntit
 	{
 		if(!tile.formed||!tile.getWorldNonnull().isBlockLoaded(tile.getPos())||tile.isDummy())
 			return;
-		final BlockRendererDispatcher blockRenderer = Minecraft.getInstance().getBlockRendererDispatcher();
-		BlockState state = tile.getWorldNonnull().getBlockState(tile.getPos());
-		if(state.getBlock()!=Multiblocks.bucketWheel)
-			return;
-		IBakedModel model = WHEEL.get(null);
 		Map<String, String> texMap = new HashMap<>();
-		List<String> list = Lists.newArrayList("bucketWheel");
+		List<String> textures = new ArrayList<>(8);
+		int offset = 0;
 		synchronized(tile.digStacks)
 		{
 			for(int i = 0; i < tile.digStacks.size(); i++)
 				if(!tile.digStacks.get(i).isEmpty())
 				{
-					list.add("dig"+i);
-					Block b = Block.getBlockFromItem(tile.digStacks.get(i).getItem());
+					offset = i;
+					break;
+				}
+			for(int i = 0; i < tile.digStacks.size(); i++)
+			{
+				int realIndex = (i+offset)%tile.digStacks.size();
+				ItemStack stackAtIndex = tile.digStacks.get(realIndex);
+				if(!stackAtIndex.isEmpty())
+				{
+					Block b = Block.getBlockFromItem(stackAtIndex.getItem());
 					BlockState digState = b!=Blocks.AIR?b.getDefaultState(): Blocks.COBBLESTONE.getDefaultState();
 					IBakedModel digModel = Minecraft.getInstance().getBlockRendererDispatcher().getBlockModelShapes().getModel(digState);
-					digModel.getParticleTexture();
-					texMap.put("dig"+i, digModel.getParticleTexture().getName().toString());
+					String texture = digModel.getParticleTexture().getName().toString();
+					texMap.put("dig"+i, texture);
+					textures.add(texture);
 				}
+				else
+					textures.add("");
+			}
 		}
-		IEObjState objState = new IEObjState(VisibilityList.show(list));
-
 		matrixStack.push();
 
 		matrixStack.translate(.5, .5, .5);
 		bufferIn = TileRenderUtils.mirror(tile, matrixStack, bufferIn);
 		float dir = tile.getFacing()==Direction.SOUTH?90: tile.getFacing()==Direction.NORTH?-90: tile.getFacing()==Direction.EAST?180: 0;
 		matrixStack.rotate(new Quaternion(new Vector3f(0, 1, 0), dir, true));
-		float rot = tile.rotation+(float)(tile.active?IEServerConfig.MACHINES.excavator_speed.get()*partialTicks: 0);
+		float rot = tile.rotation-45*offset+(float)(tile.active?IEServerConfig.MACHINES.excavator_speed.get()*partialTicks: 0);
 		matrixStack.rotate(new Quaternion(new Vector3f(1, 0, 0), rot, true));
 
 		matrixStack.translate(-.5, -.5, -.5);
-		IVertexBuilder builder = bufferIn.getBuffer(RenderType.getSolid());
-		IModelData modelData = new SinglePropertyModelData<>(tile, IOBJModelCallback.PROPERTY);
-		List<BakedQuad> quads;
-		if(model instanceof IESmartObjModel)
-			quads = ((IESmartObjModel)model).getQuads(state, objState, texMap,
-					true, modelData);
-		else
-			quads = model.getQuads(state, null, Utils.RAND, modelData);
-		ClientUtils.renderModelTESRFast(quads, builder, matrixStack, combinedLightIn);
+		try
+		{
+			CACHED_BUFFERS.get(textures, () -> IVertexBufferHolder.create(() -> {
+				IBakedModel model = WHEEL.get(null);
+				BlockState state = Multiblocks.bucketWheel.getDefaultState();
+				List<String> list = Lists.newArrayList("bucketWheel");
+				list.addAll(texMap.keySet());
+				IEObjState objState = new IEObjState(VisibilityList.show(list));
+				if(model instanceof IESmartObjModel)
+					return ((IESmartObjModel)model).getQuads(state, objState, texMap, true, EmptyModelData.INSTANCE);
+				else
+					return model.getQuads(state, null, Utils.RAND, EmptyModelData.INSTANCE);
+			})).render(RenderType.getSolid(), combinedLightIn, combinedOverlayIn, bufferIn, matrixStack);
+		} catch(ExecutionException ex)
+		{
+			throw new RuntimeException(ex);
+		}
 		matrixStack.pop();
+	}
+
+	public static void reset()
+	{
+		CACHED_BUFFERS.invalidateAll();
 	}
 }