+++ src/main/java/blusunrize/immersiveengineering/common/util/RecipeSerializers.java
@@ -96,6 +96,9 @@ public class RecipeSerializers
 		CrusherRecipe.SERIALIZER = RECIPE_SERIALIZERS.register(
 				"crusher", CrusherRecipeSerializer::new
 		);
+		SawmillRecipe.SERIALIZER = RECIPE_SERIALIZERS.register(
+				"sawmill", SawmillRecipeSerializer::new
+		);
 		FermenterRecipe.SERIALIZER = RECIPE_SERIALIZERS.register(
 				"fermenter", FermenterRecipeSerializer::new
 		);